moved data generators into a separate directory, re-added the LootTableGenerator because for me the 'connection/link' was broken, and added the VegetationGenerators class
