+++ src/main/java/luu/indepth/core/data/ModWorldGenerator.java
@@ -1,4 +1,4 @@
-package luu.indepth.data;
+package luu.indepth.core.data;
 
 import luu.indepth.Indepth;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
@@ -8,6 +8,8 @@ import net.minecraft.registry.RegistryWrapper;
 
 import java.util.concurrent.CompletableFuture;
 
+import static luu.indepth.Indepth.MOD_ID;
+
 public class ModWorldGenerator extends FabricDynamicRegistryProvider {
     public ModWorldGenerator(FabricDataOutput output, CompletableFuture<RegistryWrapper.WrapperLookup> registriesFuture) {
         super(output, registriesFuture);
@@ -16,10 +18,9 @@ public class ModWorldGenerator extends FabricDynamicRegistryProvider {
     @Override
     protected void configure(RegistryWrapper.WrapperLookup registries, Entries entries) {
         entries.addAll(registries.getWrapperOrThrow(RegistryKeys.CONFIGURED_FEATURE));
+		entries.addAll(registries.getWrapperOrThrow(RegistryKeys.PLACED_FEATURE));
     }
 
     @Override
-    public String getName() {
-        return Indepth.MOD_ID;
-    }
+    public String getName() { return MOD_ID; }
 }
